[test] fix issue4521-notification-perm: use node server for win32

+++ test/sanity/issue4521-notification-perm/test.py
@@ -28,17 +28,15 @@ chrome_options.add_argument("skip-reopen-app-devtools")
 testdir = os.path.dirname(os.path.abspath(__file__))
 os.chdir(testdir)
 
-port = str(utils.free_port())
-server_path = os.path.join(testdir, '../', 'http-server.py')
-server = subprocess.Popen(['python', server_path, port])
+try:
+    os.remove('port.txt')
+except:
+    pass
 
-tpl = open('index.tpl', 'r')
-content = tpl.read().replace('{port}', port)
-tpl.close()
+server = subprocess.Popen(['python', '../http-server-node.py', 'server.js'])
 
-html = open('index.html', 'w')
-html.write(content)
-html.close()
+while not os.path.exists('port.txt') :
+    time.sleep(1)
 
 driver = webdriver.Chrome(executable_path=os.environ['CHROMEDRIVER'], chrome_options=chrome_options, service_log_path="log", service_args=["--verbose"])
 driver.implicitly_wait(2)
@@ -123,6 +121,10 @@ try:
     # perm for remote frame
     test_perm(driver, 'show-perm-frame-remote', 'perm-frame-remote')
 finally:
-    server.terminate()
+    import platform
+    if platform.system() == 'Windows':
+        subprocess.call(['taskkill', '/F', '/T', '/PID', str(server.pid)])
+    else:
+        server.terminate()
     driver.quit()
